Core cleaning
Review naming of core files and include.
