+++ cores/arduino/Arduino.h
@@ -20,64 +20,21 @@
 #ifndef Arduino_h
 #define Arduino_h
 
-#include <stdint.h>
-#include <stdlib.h>
-#include <stdbool.h>
-#include <string.h>
-#include <math.h>
-
-#include "binary.h"
-#include "itoa.h"
-
-#ifdef __cplusplus
-extern "C"{
-#endif // __cplusplus
-
-// Includes CMSIS
-#include <chip.h>
-
-#include "wiring_constants.h"
-
-#define clockCyclesPerMicrosecond() ( SystemCoreClock / 1000000L )
-#define clockCyclesToMicroseconds(a) ( ((a) * 1000L) / (SystemCoreClock / 1000L) )
-#define microsecondsToClockCycles(a) ( (a) * (SystemCoreClock / 1000000L) )
-
-void yield(void);
+#include "wiring.h"
 
 /* sketch */
 extern void setup( void ) ;
 extern void loop( void ) ;
 
-/* Define attribute */
-#if defined   ( __CC_ARM   ) /* Keil uVision 4 */
-    #define WEAK (__attribute__ ((weak)))
-#elif defined ( __ICCARM__ ) /* IAR Ewarm 5.41+ */
-    #define WEAK __weak
-#elif defined (  __GNUC__  ) /* GCC CS */
-    #define WEAK __attribute__ ((weak))
-#endif
-
+#ifdef __cplusplus
+extern "C"{
+#endif // __cplusplus
+void yield(void);
 #ifdef __cplusplus
 } // extern "C"
 #endif // __cplusplus
 
-#ifdef __cplusplus
-#include "WCharacter.h"
-#include "WString.h"
-#include "Tone.h"
-#include "WMath.h"
-#include "HardwareSerial.h"
-#include "wiring_pulse.h"
-#endif // __cplusplus
-
-
 // Include board variant
 #include "pins_arduino.h"
-
-#include "wiring.h"
-#include "wiring_digital.h"
-#include "wiring_analog.h"
-#include "wiring_shift.h"
-#include "WInterrupts.h"
 
 #endif // Arduino_h
