+++ cores/arduino/wiring.h
@@ -17,62 +17,46 @@
   Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 */
 
-#ifndef _WIRING_
-#define _WIRING_
+#ifndef _WIRING_H_
+#define _WIRING_H_
+
+#include <stdint.h>
+#include <stdlib.h>
+#include <stdbool.h>
+#include <string.h>
+#include <math.h>
+
+#include "binary.h"
+#include "itoa.h"
 
 #ifdef __cplusplus
-extern "C" {
-#endif
-
-/**
- *
- */
-extern void initVariant( void ) ;
-extern void init( void ) ;
-
-/**
- * \brief Returns the number of milliseconds since the Arduino board began running the current program.
- *
- * This number will overflow (go back to zero), after approximately 50 days.
- *
- * \return Number of milliseconds since the program started (uint32_t)
- */
-extern uint32_t millis( void ) ;
-
-/**
- * \brief Returns the number of microseconds since the Arduino board began running the current program.
- *
- * This number will overflow (go back to zero), after approximately 70 minutes. On 16 MHz Arduino boards
- * (e.g. Duemilanove and Nano), this function has a resolution of four microseconds (i.e. the value returned is
- * always a multiple of four). On 8 MHz Arduino boards (e.g. the LilyPad), this function has a resolution
- * of eight microseconds.
- *
- * \note There are 1,000 microseconds in a millisecond and 1,000,000 microseconds in a second.
- */
-extern uint32_t micros( void ) ;
-
-/**
- * \brief Pauses the program for the amount of time (in miliseconds) specified as parameter.
- * (There are 1000 milliseconds in a second.)
- *
- * \param dwMs the number of milliseconds to pause (uint32_t)
- */
-extern void delay( uint32_t dwMs ) ;
-
-/**
- * \brief Pauses the program for the amount of time (in microseconds) specified as parameter.
- *
- * \param dwUs the number of microseconds to pause (uint32_t)
- */
-static inline void delayMicroseconds(uint32_t) __attribute__((always_inline, unused));
-static inline void delayMicroseconds(uint32_t usec){
-  uint32_t start = GetCurrentMicro();
-
-  while((start+usec) > GetCurrentMicro());
-}
-
+extern "C"{
+#endif // __cplusplus
+#include <board.h>
 #ifdef __cplusplus
 }
 #endif
 
-#endif /* _WIRING_ */
+#include "wiring_analog.h"
+#include "wiring_constants.h"
+#include "wiring_digital.h"
+#include "wiring_pulse.h"
+#include "wiring_shift.h"
+#include "wiring_time.h"
+#include "WInterrupts.h"
+
+#ifdef __cplusplus
+#include "HardwareSerial.h"
+#include "Tone.h"
+#include "WCharacter.h"
+#include "WMath.h"
+#include "WString.h"
+#endif // __cplusplus
+
+#define clockCyclesPerMicrosecond() ( SystemCoreClock / 1000000L )
+#define clockCyclesToMicroseconds(a) ( ((a) * 1000L) / (SystemCoreClock / 1000L) )
+#define microsecondsToClockCycles(a) ( (a) * (SystemCoreClock / 1000000L) )
+
+
+
+#endif /* _WIRING_H_ */
